src/02-faq-agent-langgraph/pyproject.toml
[project]
name = "02-faq-agent-langgraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.0",
    "openai>=2.1.0",
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
